+++ CHANGES.md
@@ -1,5 +1,25 @@
 twilio-go changelog
 ====================
+[2022-10-31] Version 1.1.1
+--------------------------
+**Library - Miscellaneous**
+- [PR #195](https://github.com/twilio/twilio-go/pull/195): move webhook cluster test. Thanks to [@claudiachua](https://github.com/claudiachua)!
+
+**Api**
+- Added `contentSid` and `contentVariables` to Message resource with public visibility as Beta
+- Add `UserDefinedMessageSubscription` and `UserDefinedMessage` resource
+
+**Proxy**
+- Remove FailOnParticipantConflict param from Proxy Session create and update and Proxy Participant create
+
+**Supersim**
+- Update SettingsUpdates resource to remove PackageSid
+
+**Taskrouter**
+- Add `Ordering` query parameter to Workers and TaskQueues for sorting by
+- Add `worker_sid` query param for list reservations endpoint
+
+
 [2022-10-19] Version 1.1.0
 --------------------------
 **Library - Docs**